bugprone-narrowing-conversions

+++ c2h/include/c2h/cpu_timer.h
@@ -55,7 +55,7 @@ public:
 
   void print_elapsed_seconds(const std::string& label)
   {
-    printf("%0.6f s: %s\n", this->elapsed_us() / 1000000.f, label.c_str());
+    printf("%0.6f s: %s\n", static_cast<float>(this->elapsed_us()) / 1000000.f, label.c_str());
   }
 
   void print_elapsed_seconds_and_reset(const std::string& label)
