+++ thrust/examples/discrete_voronoi.cu
@@ -199,7 +199,7 @@ void jfa(thrust::device_vector<int>& in, thrust::device_vector<int>& out, unsign
       thrust::counting_iterator<int>(0))
       + n * m, // NOLINT(bugprone-misplaced-widening-cast)
     out.begin(),
-    voronoi_site_selector(m, n, k));
+    voronoi_site_selector(m, n, static_cast<int>(k)));
 }
 /********************************************/
 
@@ -244,7 +244,7 @@ int main() // NOLINT(bugprone-exception-escape)
   }
 
   display_time(t);
-  std::cout << "  ( " << seeds.size() / (1e6 * t.elapsed()) << " MPixel/s ) " << '\n';
+  std::cout << "  ( " << static_cast<double>(seeds.size()) / (1e6 * t.elapsed()) << " MPixel/s ) " << '\n';
 
   std::cout << "[Device to Host Copy]" << '\n';
   t.restart();
